Add new base classes for actions
As it stands, there are two kinds of actions in Review Board today:

* old-style, dict based actions, which are used by extensions.
* class-based actions, where are used internally (and can experimentally
  be enabled for extension use).

This patch is the first in a series that seeks to overhaul the actions
system so that only class-based actions are used *everywhere* (and
old-style, dict-based actions are implicitly converted to class-based
actions for compatability).

This patch lays the ground work by adding a base class for all actions
and a mixin for tranforming actions into menus.