Order DiffCommits by primary key by default
There are a few places where we are querying for DiffCommits but aren't
providing an `ORDER BY` clause, so we get them in whatever order the
database provides. SQLite returns them by primary key, but MySQL does
not, which leads to commits being out of order is certain places (such
as the CommitListField in change descriptions).

We *always* want DiffCommits ordered by primary key, so we make that the
default. Any query in which this is not the case can explicitly provide
an `ORDER BY` clause.